Removed PS from maxEnvelopeSize and improved enable/disable wdigest
