+++ CIMplant/CIMplant.csproj
@@ -67,6 +67,7 @@
   <ItemGroup>
     <Compile Include="Commander.cs" />
     <Compile Include="Connector.cs" />
+    <Compile Include="EnvelopeSize.cs" />
     <Compile Include="ExceptionLogging.cs" />
     <Compile Include="ExecuteWMIwithProvider.cs" />
     <Compile Include="ExecuteCIM.cs" />
@@ -76,6 +77,7 @@
     <Compile Include="Planter.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RegistryMod.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\.editorconfig">
