+++ CIMplant/CIMplant.csproj.user
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
+    <StartArguments>-c disable_wdigest -s 192.168.64.4 -u test -p 1</StartArguments>
+  </PropertyGroup>
+</Project>
